move common to libs
